pyproject.toml
[tool.poetry]
name = "twitter-clone"
version = "0.1.0"
description = ""
authors = ["pedrosimiao <pedrolumiere@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-debug-toolbar = "^5.0.1"
django-extensions = "^3.2.3"
djangorestframework-simplejwt = "^5.4.0"
pillow = "^11.1.0"
factory-boy = "^3.3.3"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
